gitlab/.env がなくても vagrant destroy ができるようにする

+++ gitlab/config.rb
@@ -2,6 +2,9 @@ require 'erb'
 require 'dotenv'
 
 # Create docker-compose.yml from erb template
-Dotenv.load! File.join(File.dirname(__FILE__), ".env")
-erb = File.open(File.join(File.dirname(__FILE__), "docker-compose.yml.erb")) { |f| ERB.new(f.read) }
-File.write(File.join(File.dirname(__FILE__), "docker-compose.yml"), erb.result(binding))
+DOT_ENV = File.join(File.dirname(__FILE__), ".env")
+if File.exist?(DOT_ENV)
+  Dotenv.load! DOT_ENV
+  erb = File.open(File.join(File.dirname(__FILE__), "docker-compose.yml.erb")) { |f| ERB.new(f.read) }
+  File.write(File.join(File.dirname(__FILE__), "docker-compose.yml"), erb.result(binding))
+end
